fix: resolve API inconsistencies and UdsServer concurrency bugs
- Unify exception type in multi_client() to BuilderException
- Fix UdsServer::max_clients() using shared_lock instead of unique_lock
- Add client_limit_enabled_ tracking to UdsServer for consistency with TcpServer
- Reset work_guard_ before ioc->stop() in UdsServer::stop()
- Change use_external_context_ and manage_external_context_ to std::atomic<bool>
- Unify UdpServerBuilder::max_clients() parameter type to uint32_t
- Fix LineFramer::max_length to be content-based (delimiter excluded)

+++ unilink/builder/udp_builder.cc
@@ -226,7 +226,7 @@ UdpServerBuilder<State>& UdpServerBuilder<State>::bind_address(const std::string
 }
 
 template <uint32_t State>
-UdpServerBuilder<State>& UdpServerBuilder<State>::max_clients(size_t max) {
+UdpServerBuilder<State>& UdpServerBuilder<State>::max_clients(uint32_t max) {
   max_clients_ = max;
   client_limit_enabled_ = true;
   return *this;
